requirements.txt
# BrainBot Dependencies
# ===================
# Core AI and UI framework dependencies for the BrainBot application

# ============ Existing Chat UI Dependencies ============
# Textual - Modern Terminal User Interface framework
textual==0.82.0

# Llama C++ Python bindings for local AI inference
llama-cpp-python==0.3.1

# Hugging Face Hub for downloading AI models
huggingface-hub==0.25.2

# Rich - Beautiful terminal formatting
rich==13.9.4

# Pyttsx3 - Text-to-speech for reading responses aloud
pyttsx3==2.90


# ============ Voice Mode Dependencies ============

# Porcupine - Wake word detection ("Hey Computer")
pvporcupine==3.0.0

# PyAudio - Audio input/output for recording and playback
pyaudio==0.2.14

# Python-dotenv - Environment variable management
python-dotenv==1.0.0


# ============ Daemon Dependencies ============

# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0

# MCP Server
fastmcp>=2.0.0

# System monitoring
psutil>=5.9.0

# Image processing (for LCDs)
Pillow>=10.0.0


# ============ Raspberry Pi Hardware (Optional) ============
# These are only needed when running on Raspberry Pi with hardware

# Uncomment on Raspberry Pi:
# RPi.GPIO>=0.7.0
# gpiozero>=2.0.0
# adafruit-circuitpython-ssd1306>=2.12.0
# adafruit-circuitpython-neopixel>=6.3.0
# board
# busio
# spidev


# ============ Future Dependencies ============
# Uncomment when implementing these features:

# Discord bot
# discord.py>=2.3.0

# YouTube integration
# google-api-python-client>=2.100.0
